+++ docs/changelog.md
@@ -1,5 +1,21 @@
 # TBD
 
+# 0.4.2
+### Features
+* Added generation of Besu genesis file
+* Added Besu EL
+
+### Fixes
+* Fixed an issue where the CL REST client would try to deserialize the bodies of responses that came back with non-200 status codes
+* When a Teku node is present, require merge fork epoch to be >= 3 as a workaround for a bug in Teku
+* Disallow a Prysm node being a boot node due to https://github.com/kurtosis-tech/eth2-merge-kurtosis-module/pull/36
+
+### Changes
+* Set the `mergeForkBlock` parameter in the EL genesis config template to `10` per Pari's recommendation
+* Switch back to [the default genesis generator](https://github.com/skylenet/ethereum-genesis-generator) (rather than the Kurtosis fork of it)
+* Nethermind genesis JSON is generated using the genesis generator image
+* Centralized EL client availability-waiter and mining-waiter logic
+
 # 0.4.1
 ### Fixes
 * Fixed an issue where using emptystring as the default image wasn't working